Move MatchingComparisonOutput to tests/types folder
